+++ flutter_local_notifications/lib/src/platform_specifics/ios/method_channel_mappers.dart
@@ -25,6 +25,7 @@ extension IOSNotificationDetailsMapper on IOSNotificationDetails {
         'presentAlert': presentAlert,
         'presentSound': presentSound,
         'presentBadge': presentBadge,
+        'subtitle': subtitle,
         'sound': sound,
         'badgeNumber': badgeNumber,
         'attachments': attachments